bot/requirements.txt
discord.py
questionary
# Styling
pyfiglet
colorama